* commit '5d6332e3744e37ced079e5360f1fa3e0e5681d57':
  AArch64: Use long for pointers in MediaDrm and MediaMuxer

+++ media/java/android/media/MediaDrm.java
@@ -100,7 +100,7 @@ public final class MediaDrm {
     private EventHandler mEventHandler;
     private OnEventListener mOnEventListener;
 
-    private int mNativeContext;
+    private long mNativeContext;
 
     /**
      * Query if the given scheme identified by its UUID is supported on
